Move ToastNotification sample to separate file
#147
Adding consistency with LiveTile sample
