+++ templates/ConsumerFeatures/ToastNotifications/Services/ActivationService_postaction.cs
@@ -9,6 +9,7 @@ namespace ItemNamespace.Services
         {
             Singleton<ToastNotificationsService>.Instance.ShowToastNotificationSample();
         }
+
         private IEnumerable<ActivationHandler> GetActivationHandlers()
         {
             yield return Singleton<ToastNotificationsService>.Instance;
