+++ templates/ConsumerFeatures/ToastNotifications/Services/ToastNotificationsFeatureService.Samples.cs
@@ -1,0 +1,64 @@
+using Microsoft.Toolkit.Uwp.Notifications;
+using Windows.UI.Notifications;
+
+namespace RootNamespace.Services
+{
+    internal partial class ToastNotificationsFeatureService
+    {
+        public void ShowToastNotificationSample()
+        {            
+            // Create the toast content
+            var content = new ToastContent()
+            {
+                // TODO UWPTemplates: Check this documentation to know more about the Launch property
+                // Documentation: https://developer.microsoft.com/en-us/windows/uwp-community-toolkit/api/microsoft_toolkit_uwp_notifications_toastcontent                
+                Launch = "ToastContentActivationParams",
+
+                Visual = new ToastVisual()
+                {
+                    BindingGeneric = new ToastBindingGeneric()
+                    {
+                        Children =
+                        {
+                            new AdaptiveText()
+                            {
+                                Text = "Action - text"
+                            },
+
+                            new AdaptiveText()
+                            {
+                                Text = @"Make sure the left button on the toast has the text ""Ok"" on it, and the right button has the text ""Cancel"" on it."
+                            }
+                        }
+                    }
+                },
+
+                Actions = new ToastActionsCustom()
+                {
+                    Buttons =
+                    {
+                        // TODO UWPTemplates: Check this documentation to know more about Toast Buttons
+                        // Documentation: https://developer.microsoft.com/en-us/windows/uwp-community-toolkit/api/microsoft_toolkit_uwp_notifications_toastbutton
+                        new ToastButton("Ok", "ToastButtonActivationArguments")
+                        {
+                            ActivationType = ToastActivationType.Foreground
+                        },
+
+                        new ToastButtonDismiss("Cancel")
+                    }
+                }
+            };
+
+            // Create the toast
+            var toast = new ToastNotification(content.GetXml())
+            {
+                // TODO UWPTemplates: Gets or sets the unique identifier of this notification within the notification Group. Max length 16 characters.
+                // Documentation: https://docs.microsoft.com/uwp/api/windows.ui.notifications.toastnotification
+                Tag = "ToastTag"
+            };
+
+            // And show the toast
+            ShowToastNotification(toast);
+        }
+    }
+}
