+++ templates/ConsumerFeatures/ToastNotifications/Services/ToastNotificationsFeatureService.cs
@@ -1,4 +1,3 @@
-using Microsoft.Toolkit.Uwp.Notifications;
 using Windows.ApplicationModel.Activation;
 using Windows.UI.Notifications;
 
@@ -8,7 +7,7 @@ using RootNamespace.Activation;
 
 namespace RootNamespace.Services
 {
-    internal class ToastNotificationsFeatureService : ActivationHandler<ToastNotificationActivatedEventArgs>
+    internal partial class ToastNotificationsFeatureService : ActivationHandler<ToastNotificationActivatedEventArgs>
     {
         public void ShowToastNotification(ToastNotification toastNotification)
         {
@@ -22,62 +21,6 @@ namespace RootNamespace.Services
             // Documentation: https://blogs.msdn.microsoft.com/tiles_and_toasts/2015/07/08/quickstart-sending-a-local-toast-notification-and-handling-activations-from-it-windows-10/
 
             await Task.CompletedTask;
-        }
-
-        public void ShowToastNotificationSample()
-        {            
-            // Create the toast content
-            var content = new ToastContent()
-            {
-                // TODO UWPTemplates: Check this documentation to know more about the Launch property
-                // Documentation: https://developer.microsoft.com/en-us/windows/uwp-community-toolkit/api/microsoft_toolkit_uwp_notifications_toastcontent                
-                Launch = "ToastContentActivationParams",
-
-                Visual = new ToastVisual()
-                {
-                    BindingGeneric = new ToastBindingGeneric()
-                    {
-                        Children =
-                        {
-                            new AdaptiveText()
-                            {
-                                Text = "Action - text"
-                            },
-
-                            new AdaptiveText()
-                            {
-                                Text = @"Make sure the left button on the toast has the text ""Ok"" on it, and the right button has the text ""Cancel"" on it."
-                            }
-                        }
-                    }
-                },
-
-                Actions = new ToastActionsCustom()
-                {
-                    Buttons =
-                    {
-                        // TODO UWPTemplates: Check this documentation to know more about Toast Buttons
-                        // Documentation: https://developer.microsoft.com/en-us/windows/uwp-community-toolkit/api/microsoft_toolkit_uwp_notifications_toastbutton
-                        new ToastButton("Ok", "ToastButtonActivationArguments")
-                        {
-                            ActivationType = ToastActivationType.Foreground
-                        },
-
-                        new ToastButtonDismiss("Cancel")
-                    }
-                }
-            };
-
-            // Create the toast
-            var toast = new ToastNotification(content.GetXml())
-            {
-                // TODO UWPTemplates: Gets or sets the unique identifier of this notification within the notification Group. Max length 16 characters.
-                // Documentation: https://docs.microsoft.com/uwp/api/windows.ui.notifications.toastnotification
-                Tag = "ToastTag"
-            };
-
-            // And show the toast
-            ShowToastNotification(toast);
         }
     }
 }
